move the evqe gate and circuit layer code to its own package
