+++ queasars/minimum_eigensolvers/evqe/quantum_circuit/__init__.py
@@ -1,0 +1,2 @@
+# Quantum Evolving Ansatz Variational Solver (QUEASARS)
+# Copyright 2024 DLR - Deutsches Zentrum für Luft- und Raumfahrt e.V.
